tests: added service tests (#8)
* tests: added oxygen-supply, room controller tests

* tests: added  room controller tests

* tests: added controller tests

* tests: added service tests

---------

+++ src/main/kotlin/ru/itmo/highload_systems/api/dto/RoomResponse.kt
@@ -8,7 +8,7 @@ class RoomResponse(
     val number: Long,
     val size: Long,
     val capacity: Long,
-    val avgPersonNorm: Float,
+    val avgPersonNorm: Double,
     var createdAt: OffsetDateTime?,
     var updatedAt: OffsetDateTime?
 )
